refactor: move packages to android
