+++ android/beagle/src/main/java/br/com/zup/beagle/android/engine/renderer/LayoutViewRendererFactory.kt
@@ -28,7 +28,7 @@ import br.com.zup.beagle.android.engine.renderer.layout.ScrollViewRenderer
 import br.com.zup.beagle.android.engine.renderer.layout.SpacerViewRenderer
 import br.com.zup.beagle.android.engine.renderer.layout.StackViewRenderer
 import br.com.zup.beagle.android.engine.renderer.layout.VerticalViewRender
-import br.com.zup.beagle.widget.layout.ScreenComponent
+import br.com.zup.beagle.android.widget.layout.ScreenComponent
 import br.com.zup.beagle.widget.form.Form
 import br.com.zup.beagle.widget.layout.Container
 import br.com.zup.beagle.widget.layout.Horizontal
