+++ android/beagle/src/main/java/br/com/zup/beagle/android/view/BeagleActivity.kt
@@ -35,7 +35,7 @@ import br.com.zup.beagle.android.setup.BeagleEnvironment
 import br.com.zup.beagle.android.utils.configureSupportActionBar
 import br.com.zup.beagle.android.utils.toComponent
 import br.com.zup.beagle.widget.layout.Screen
-import br.com.zup.beagle.widget.layout.ScreenComponent
+import br.com.zup.beagle.android.widget.layout.ScreenComponent
 import kotlinx.android.parcel.Parcelize
 
 sealed class ServerDrivenState {
