+++ android/beagle/src/main/java/br/com/zup/beagle/android/widget/layout/ScreenComponent.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package br.com.zup.beagle.widget.layout
+package br.com.zup.beagle.android.widget.layout
 
 import br.com.zup.beagle.analytics.ScreenAnalytics
 import br.com.zup.beagle.analytics.ScreenEvent
@@ -22,6 +22,7 @@ import br.com.zup.beagle.core.Appearance
 import br.com.zup.beagle.core.AppearanceComponent
 import br.com.zup.beagle.core.LayoutComponent
 import br.com.zup.beagle.core.ServerDrivenComponent
+import br.com.zup.beagle.widget.layout.NavigationBar
 
 internal data class ScreenComponent(
     val identifier: String? = null,
