+++ android/beagle/src/main/java/br/com/zup/beagle/android/widget/ui/UndefinedWidget.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package br.com.zup.beagle.widget.ui
+package br.com.zup.beagle.android.widget.ui
 
 import android.content.Context
 import android.view.View
