+++ android/beagle/src/test/java/br/com/zup/beagle/android/engine/renderer/ui/UndefinedViewRendererTest.kt
@@ -26,7 +26,7 @@ import br.com.zup.beagle.android.setup.BeagleEnvironment
 import br.com.zup.beagle.android.setup.Environment
 import br.com.zup.beagle.android.view.BeagleTextView
 import br.com.zup.beagle.android.view.ViewFactory
-import br.com.zup.beagle.widget.ui.UndefinedWidget
+import br.com.zup.beagle.android.widget.ui.UndefinedWidget
 import io.mockk.Runs
 import io.mockk.every
 import io.mockk.impl.annotations.InjectMockKs
