+++ android/beagle/src/test/java/br/com/zup/beagle/android/utils/ToolbarManagerTest.kt
@@ -35,7 +35,7 @@ import br.com.zup.beagle.android.testutil.RandomData
 import br.com.zup.beagle.android.view.BeagleActivity
 import br.com.zup.beagle.widget.layout.NavigationBar
 import br.com.zup.beagle.widget.layout.NavigationBarItem
-import br.com.zup.beagle.widget.layout.ScreenComponent
+import br.com.zup.beagle.android.widget.layout.ScreenComponent
 import io.mockk.Runs
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
